requirements.txt
asgiref==3.11.1
Django==6.0.4
django-environ==0.13.0
django-ratelimit==4.1.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.15.0
gunicorn==25.3.0
packaging==26.1
psycopg2-binary==2.9.12
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2026.1
